Moving tests into cookiecutter scheme
